requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.28.1
supabase==2.15.2
python-dotenv==1.0.0
openai==1.14.0
python-multipart==0.0.6
pydantic==2.5.0
typing-extensions==4.8.0
Pillow==10.1.0
volcengine-python-sdk==3.0.9
requests==2.31.0 